backend/requirements.txt
flask
pymongo
python-jose
requests
fastapi